refactor(product): respect complexity budgets
