+++ framework/tui/src/agent-work-lab-view.tsx
@@ -1,16 +1,23 @@
 // SPDX-License-Identifier: Apache-2.0
 
+import fs from 'node:fs';
+import path from 'node:path';
 import type {
   AgentWorkLab,
   AgentWorkLabEvent,
   AgentWorkLabReport,
   AgentWorkLabStartupRoute,
+  KungfuOnboardingState,
   ProjectTemplateCreationReceipt,
   ProjectTemplatePlan,
   ProjectTemplateWorkspaceSelection,
 } from '@kungfu-tech/api/capability';
-import { agentWorkLabRunProgressLabel } from '@kungfu-tech/api/capability';
-import { useApp } from 'ink';
+import {
+  DEFAULT_KUNGFU_ONBOARDING_STATE,
+  agentWorkLabRunProgressLabel,
+  parseKungfuOnboardingState,
+} from '@kungfu-tech/api/capability';
+import { Box, Text, useApp } from 'ink';
 import React from 'react';
 import {
   AGENT_WORK_LAB_CHECKS,
@@ -24,6 +31,7 @@ import {
   type QuickCommand,
   SessionWorkbench,
   type TerminalDimensions,
+  TitledBorderWindow,
   type WorkbenchActionButton,
   type WorkbenchCheck,
   type WorkbenchFocus,
@@ -48,6 +56,129 @@ import {
   workbenchViewportRows,
 } from './profile-shell.js';
 import { decodeTerminalMouseInput } from './terminal-lifecycle.js';
+
+export function readTuiOnboardingState(
+  configHome: string,
+  readFile: (file: string) => string = (file) => fs.readFileSync(file, 'utf8'),
+): KungfuOnboardingState {
+  try {
+    const value = JSON.parse(
+      readFile(path.join(configHome, 'config.json')),
+    ) as { ui?: { onboarding?: unknown } };
+    return parseKungfuOnboardingState(value.ui?.onboarding);
+  } catch {
+    return { ...DEFAULT_KUNGFU_ONBOARDING_STATE };
+  }
+}
+
+function wrapOnboardingText(value: string, width: number): string[] {
+  const words = value
+    .split(/\s+/u)
+    .flatMap((word) =>
+      word.length <= width
+        ? [word]
+        : Array.from({ length: Math.ceil(word.length / width) }, (_, index) =>
+            word.slice(index * width, (index + 1) * width),
+          ),
+    );
+  const lines: string[] = [];
+  let line = '';
+  for (const word of words) {
+    if (!line) line = word;
+    else if (`${line} ${word}`.length <= width) line = `${line} ${word}`;
+    else {
+      lines.push(line);
+      line = word;
+    }
+  }
+  if (line) lines.push(line);
+  return lines;
+}
+
+export type TuiOnboardingAction =
+  | 'agent'
+  | 'lab'
+  | 'tour'
+  | 'continue'
+  | 'dismiss';
+
+export function AgentFirstOnboardingView({
+  dimensions,
+  state,
+  command,
+  prompt,
+  notice = '',
+  onAction,
+}: {
+  dimensions: TerminalDimensions;
+  state: KungfuOnboardingState;
+  command: string;
+  prompt: string;
+  notice?: string;
+  onAction: (action: TuiOnboardingAction) => void;
+}) {
+  const { exit } = useApp();
+  React.useEffect(() => {
+    const onData = (chunk: Buffer | string) => {
+      const value = String(chunk);
+      if (value === 'q' || value === 'Q' || value === '\u0003') exit();
+      else if (value === '1' || value === '\r' || value === '\n')
+        onAction('agent');
+      else if (value === '2') onAction('lab');
+      else if (value === '3') onAction('tour');
+      else if (value === 'w') onAction('continue');
+      else if (value === 's') onAction('dismiss');
+    };
+    process.stdin.on('data', onData);
+    return () => {
+      process.stdin.off('data', onData);
+    };
+  }, [exit, onAction]);
+
+  return (
+    <Box
+      width={dimensions.columns}
+      height={dimensions.rows}
+      paddingX={1}
+      flexDirection="column"
+      overflow="hidden"
+    >
+      <Text color="green" bold>
+        KUNGFU · AGENT-FIRST ENTRY
+      </Text>
+      <Text bold>Keep your agent. Give it durable Work.</Text>
+      <Text wrap="wrap">
+        Kungfu does not require a new chat or daily workspace. Start by teaching
+        the agent you already use how to preserve Projects, Work, attempts,
+        review, and settlement across sessions.
+      </Text>
+      <TitledBorderWindow
+        columns={Math.max(20, dimensions.columns - 2)}
+        title="1 · COPY THIS TO YOUR EXISTING AGENT"
+        borderColor="green"
+        paddingX={1}
+        rows={[
+          ...wrapOnboardingText(prompt, Math.max(16, dimensions.columns - 8)),
+          '',
+          ...wrapOnboardingText(
+            `Exact local command: ${command}`,
+            Math.max(16, dimensions.columns - 8),
+          ),
+          '[1/Enter] I’ll use my Agent',
+        ]}
+      />
+      <Text>
+        Optional: [2] Agent Work Lab · [3] Guided Project Tour · [w] Work now ·
+        [s] Don’t show again
+      </Text>
+      <Text dimColor>
+        Keep using kungfu run codex|claude|opencode|amp. GUI/TUI are optional
+        control surfaces. Current route: {state.route}.
+      </Text>
+      {notice ? <Text color="yellow">{notice}</Text> : null}
+    </Box>
+  );
+}
 
 export type TuiAgentWorkLabMode = AgentWorkLabCaseId;
 export type TuiAgentWorkLabFocus = WorkbenchFocus;
